Edits the Ship class to begin refactoring the ship to not rely on a direct reference to the screen.
The screen should not be in the objects of the game I think it makes more sense to have objects we can pass into other objects to represent the parts of the gameScreen we actually need. This is critical for testing purposes, if our objects require a working reference to the gameScreen then we have to work really hard to unit test anything because the work will depend on the screen's realtime execution. Instead we can mock the objects and have better object hygiene