Fixed tests to use managers
This is a big deal. Reminder, tests are meant to be in stone, and we are changing them. This should only be done if it's ABSOLUTELY NECESSARY. It is here, as we are changing how powerful tiles are, and we want all tiles to be in possession of a self-destruct button in the form of their managers.